backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "repoarkeolog-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=12.0",
    "python-multipart>=0.0.9",
    # Task queue
    "celery>=5.4.0",
    "redis>=5.0.4",
    # Database
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",
    # Pydantic / settings
    "pydantic>=2.7.1",
    "pydantic-settings>=2.3.0",
    # AI / LLM
    "google-generativeai>=0.7.2",
    "groq>=0.9.0",
    # Vector DB
    "qdrant-client>=1.9.1",
    # Git & code analysis
    "gitpython>=3.1.43",
    "tree-sitter>=0.22.3",
    "tree-sitter-python>=0.22.0",
    "tree-sitter-javascript>=0.22.0",
    "tree-sitter-typescript>=0.22.0",
    # Static analysis
    "ruff>=0.4.7",
    "radon>=6.0.1",
    "vulture>=2.11",
    # Utilities
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "networkx>=3.3",
    "aiofiles>=23.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
]
prod = []

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
